Move hosts tests into its own file hosts_test.go
